+++ src/helpers/index.ts
@@ -1,2 +1,1 @@
-export { default as Environment } from './Environment';
 export { default as Logger } from './Logger';