+++ CHANGELOG.md
@@ -11,3 +11,10 @@
 ## 1.1.0 at (`2023-10-28`)
 
 * Dependencies update.
+
+## 2.0.0 at (`2023-10-29`)
+
+* Improvements types definitions to make it easier to use;
+* Adds `HttpInsecureServer`, `HttpSecureServer`, `Http2InsecureServer` e `Http2SecureServer` classes;
+* Removes `ApiServer` class;
+* Adds `./samples` folder.